remove `javax.annotation.security` mappers
