+++ security-annotations/src/main/java/io/micronaut/security/annotation/JakartaPermitAllAnnotationMapper.java
@@ -15,7 +15,14 @@
  */
 package io.micronaut.security.annotation;
 
+import io.micronaut.core.annotation.AnnotationValue;
 import io.micronaut.core.annotation.Internal;
+import io.micronaut.inject.annotation.NamedAnnotationMapper;
+import io.micronaut.inject.visitor.VisitorContext;
+
+import java.lang.annotation.Annotation;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * Allows using the {@link jakarta.annotation.security.PermitAll} annotation in Micronaut.
@@ -23,9 +30,20 @@ import io.micronaut.core.annotation.Internal;
  * @author Fredrik Hov
  */
 @Internal
-public class JakartaPermitAllAnnotationMapper extends PermitAllAnnotationMapper {
+public class JakartaPermitAllAnnotationMapper implements NamedAnnotationMapper {
     @Override
     public String getName() {
         return "jakarta.annotation.security.PermitAll";
+    }
+
+    @Override
+    public List<AnnotationValue<?>> map(AnnotationValue<Annotation> annotation, VisitorContext visitorContext) {
+        List<AnnotationValue<?>> annotationValues = new ArrayList<>(1);
+        annotationValues.add(
+            AnnotationValue.builder(Secured.class)
+                .value("isAnonymous()")
+                .build()
+        );
+        return annotationValues;
     }
 }
