+++ security-annotations/src/main/java/io/micronaut/security/annotation/JakartaRolesAllowedAnnotationMapper.java
@@ -15,7 +15,14 @@
  */
 package io.micronaut.security.annotation;
 
+import io.micronaut.core.annotation.AnnotationValue;
 import io.micronaut.core.annotation.Internal;
+import io.micronaut.inject.annotation.NamedAnnotationMapper;
+import io.micronaut.inject.visitor.VisitorContext;
+
+import java.lang.annotation.Annotation;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * Allows using the {@link jakarta.annotation.security.RolesAllowed} annotation in Micronaut.
@@ -23,9 +30,22 @@ import io.micronaut.core.annotation.Internal;
  * @author Fredrik Hov
  */
 @Internal
-public class JakartaRolesAllowedAnnotationMapper extends RolesAllowedAnnotationMapper {
+public class JakartaRolesAllowedAnnotationMapper implements NamedAnnotationMapper {
     @Override
     public String getName() {
         return "jakarta.annotation.security.RolesAllowed";
+    }
+
+    @Override
+    public List<AnnotationValue<?>> map(AnnotationValue<Annotation> annotation, VisitorContext visitorContext) {
+        String[] values = annotation.get("value", String[].class).orElse(new String[0]);
+
+        List<AnnotationValue<?>> annotationValues = new ArrayList<>(1);
+        annotationValues.add(
+            AnnotationValue.builder(Secured.class)
+                .values(values)
+                .build()
+        );
+        return annotationValues;
     }
 }
