+++ security-annotations/src/main/resources/META-INF/services/io.micronaut.inject.annotation.AnnotationMapper
@@ -1,6 +1,3 @@
-io.micronaut.security.annotation.PermitAllAnnotationMapper
 io.micronaut.security.annotation.JakartaPermitAllAnnotationMapper
-io.micronaut.security.annotation.RolesAllowedAnnotationMapper
 io.micronaut.security.annotation.JakartaRolesAllowedAnnotationMapper
-io.micronaut.security.annotation.DenyAllAnnotationMapper
 io.micronaut.security.annotation.JakartaDenyAllAnnotationMapper
